refactor: format utils 분리, 테스트코드 path alias 적용
