+++ src/refactoring/helpers/cart.ts
@@ -2,10 +2,6 @@ import { CartItem, Coupon, Product } from '@/types';
 
 import { getMaxApplicableDiscount } from './discount';
 
-export const formatCurrency = (price: number) => {
-  return price.toLocaleString();
-};
-
 export const getRemainingStock = (cart: CartItem[], product: Product) => {
   const cartItem = cart.find((item) => item.product.id === product.id);
   return product.stock - (cartItem?.quantity || 0);
