+++ src/refactoring/helpers/coupon.ts
@@ -1,5 +1,5 @@
 import { Coupon } from '@/types';
 
-export const formatDiscountValue = (coupon: Coupon) => {
+export const formatCouponDiscountValue = (coupon: Coupon) => {
   return coupon.discountType === 'amount' ? `${coupon.discountValue}원` : `${coupon.discountValue}%`;
 };
