+++ src/refactoring/hooks/index.ts
@@ -3,3 +3,4 @@ export * from './admin/useEditingProduct.ts';
 export * from './admin/useNewCoupon.ts';
 export * from './admin/useNewDiscount.ts';
 export * from './admin/useNewProduct.ts';
+export * from './admin/useOpenProductIds.ts';
